+++ src/main/java/com/celements/payment/container/EncryptedComputopData.java
@@ -1,9 +1,10 @@
 package com.celements.payment.container;
 
+import javax.annotation.Nullable;
 import javax.annotation.concurrent.Immutable;
 import javax.validation.constraints.NotNull;
 
-import com.google.common.base.Preconditions;
+import com.google.common.base.Optional;
 
 @Immutable
 public class EncryptedComputopData {
@@ -11,14 +12,13 @@ public class EncryptedComputopData {
   private final String cipherText;
   private final int plainDataLength;
 
-  public EncryptedComputopData(@NotNull String cipherText, int plainDataLength) {
-    Preconditions.checkNotNull(cipherText);
+  public EncryptedComputopData(@Nullable String cipherText, int plainDataLength) {
     this.cipherText = cipherText;
     this.plainDataLength = plainDataLength;
   }
 
-  public @NotNull String getCipherText() {
-    return cipherText;
+  public @NotNull Optional<String> getCipherText() {
+    return Optional.fromNullable(cipherText);
   }
 
   public int getPlainDataLength() {
@@ -26,7 +26,7 @@ public class EncryptedComputopData {
   }
 
   public @NotNull String getData() {
-    return getCipherText();
+    return getCipherText().or("");
   }
 
   public int getLen() {
