+++ src/main/java/com/celements/payment/exception/ComputopCryptoException.java
@@ -4,6 +4,10 @@ public class ComputopCryptoException extends Exception {
 
   private static final long serialVersionUID = -5534681246083544419L;
 
+  public ComputopCryptoException(String msg) {
+    super(msg);
+  }
+
   public ComputopCryptoException(String msg, Exception excp) {
     super(msg, excp);
   }
