+++ src/main/java/com/celements/payment/service/ComputopService.java
@@ -59,6 +59,37 @@ public class ComputopService implements ComputopServiceRole {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ComputopService.class);
 
+  static final String MERCHANT_ID_PROP = "computop_merchant_id";
+
+  static final String BLOWFISH = "Blowfish";
+  static final String BLOWFISH_ECB_PADDED = BLOWFISH + "/ECB/PKCS5Padding";
+  static final String BLOWFISH_ECB_UNPADDED = BLOWFISH + "/ECB/NoPadding";
+  static final String BLOWFISH_SECRET_KEY_PROP = "computop_blowfish_secret_key";
+
+  static final String HMAC_SHA256 = "HmacSHA256";
+  static final String HMAC_SECRET_KEY_PROP = "computop_hmac_secret_key";
+
+  enum ReturnUrl {
+    SUCCESS("computop_return_url_success", "URLSuccess"), FAILURE("computop_return_url_failure",
+        "URLFailure"), CALLBACK("computop_return_url_callback", "URLNotify");
+
+    private final String value;
+    private final String param;
+
+    private ReturnUrl(String value, String param) {
+      this.value = value;
+      this.param = param;
+    }
+
+    public @NotNull String getValue() {
+      return value;
+    }
+
+    public @NotNull String getParamName() {
+      return param;
+    }
+  }
+
   @Requirement
   private ConfigurationSource configSrc;
 
@@ -163,18 +194,22 @@ public class ComputopService implements ComputopServiceRole {
 
   byte[] decryptString(EncryptedComputopData encryptedCallback, final SecretKey key)
       throws ComputopCryptoException {
-    CharSequence cs = encryptedCallback.getCipherText().toUpperCase();
-    LOGGER.debug("decrypting cipher [{}]", cs);
-    byte[] decodedCipher = BaseEncoding.base16().decode(cs);
-    try {
-      Cipher cipher = getCipher(Cipher.DECRYPT_MODE, BLOWFISH_ECB_UNPADDED, key);
-      int len = encryptedCallback.getPlainDataLength();
-      byte[] unpadded = cipher.doFinal(decodedCipher);
-      byte[] deciphered = Arrays.copyOfRange(unpadded, 0, min(max(len, 0), unpadded.length));
-      LOGGER.debug("decryped plain [{}]", new String(deciphered));
-      return deciphered;
-    } catch (IllegalBlockSizeException | BadPaddingException excp) {
-      throw new ComputopCryptoException("Exception decrypting message", excp);
+    if (encryptedCallback.getCipherText().isPresent()) {
+      CharSequence cs = encryptedCallback.getCipherText().get().toUpperCase();
+      LOGGER.debug("decrypting cipher [{}]", cs);
+      byte[] decodedCipher = BaseEncoding.base16().decode(cs);
+      try {
+        Cipher cipher = getCipher(Cipher.DECRYPT_MODE, BLOWFISH_ECB_UNPADDED, key);
+        int len = encryptedCallback.getPlainDataLength();
+        byte[] unpadded = cipher.doFinal(decodedCipher);
+        byte[] deciphered = Arrays.copyOfRange(unpadded, 0, min(max(len, 0), unpadded.length));
+        LOGGER.debug("decryped plain [{}]", new String(deciphered));
+        return deciphered;
+      } catch (IllegalBlockSizeException | BadPaddingException excp) {
+        throw new ComputopCryptoException("Exception decrypting message", excp);
+      }
+    } else {
+      throw new ComputopCryptoException("Ciphertext to decrypted is absent");
     }
   }
