+++ src/test/java/com/celements/payment/service/ComputopServiceTest.java
@@ -21,6 +21,7 @@ package com.celements.payment.service;
  */
 
 import static com.celements.common.test.CelementsTestUtils.*;
+import static com.celements.payment.service.ComputopService.*;
 import static com.celements.payment.service.ComputopServiceRole.*;
 import static org.easymock.EasyMock.*;
 import static org.junit.Assert.*;
@@ -39,7 +40,7 @@ import org.xwiki.configuration.ConfigurationSource;
 import com.celements.common.test.AbstractComponentTest;
 import com.celements.payment.container.EncryptedComputopData;
 import com.celements.payment.exception.ComputopCryptoException;
-import com.celements.payment.service.ComputopServiceRole.ReturnUrl;
+import com.celements.payment.service.ComputopService.ReturnUrl;
 import com.xpn.xwiki.web.Utils;
 
 public class ComputopServiceTest extends AbstractComponentTest {
